# packages/contract/openapi.yaml
components:
  schemas:
    github_com_bingoring_forin_server_internal_curriculum.BuildingGroup:
      properties:
        building:
          type: string
        floors:
          items:
            $ref: '#/components/schemas/github_com_bingoring_forin_server_internal_curriculum.FloorGroup'
          type: array
          uniqueItems: false
      type: object
    github_com_bingoring_forin_server_internal_curriculum.CurriculumState:
      properties:
        building:
          type: string
        done:
          type: integer
        floor:
          type: string
        key:
          type: string
        name:
          type: string
        next:
          description: name of the current (now) step
          type: string
        resume:
          type: boolean
        state:
          description: done | doing | todo
          type: string
        steps:
          items:
            $ref: '#/components/schemas/github_com_bingoring_forin_server_internal_curriculum.StepState'
          type: array
          uniqueItems: false
        total:
          type: integer
        where:
          type: string
      type: object
    github_com_bingoring_forin_server_internal_curriculum.FloorGroup:
      properties:
        curricula:
          items:
            $ref: '#/components/schemas/github_com_bingoring_forin_server_internal_curriculum.CurriculumState'
          type: array
          uniqueItems: false
        floor:
          type: string
        where:
          type: string
      type: object
    github_com_bingoring_forin_server_internal_curriculum.StepState:
      properties:
        kind:
          type: string
        name:
          type: string
        optional:
          description: bonus practice; doesn't gate
          type: boolean
        scenarioId:
          type: string
        state:
          description: done | now | lock | optional
          type: string
      type: object
...
    github_com_bingoring_forin_server_internal_domain_progress.GrowthStats:
      properties:
        activeDates:
          description: yyyy-mm-dd (tz), the StreakWindowDays ending today
          items:
            type: string
          type: array
          uniqueItems: false
        conversationSecondsToday:
          type: integer
        conversationSecondsWeek:
          type: integer
        newCardsToday:
          type: integer
        newCardsWeek:
          type: integer
        scenariosToday:
          type: integer
        scenariosTotal:
          description: lifetime clears → praise stickers (1 each)
          type: integer
        scenariosWeek:
          type: integer
      type: object
    github_com_bingoring_forin_server_internal_domain_progress.Progress:
      properties:
        level:
          type: integer
        rank:
          type: string
        reputation:
          description: |-
            Reputation is ordered for display and carries its own labels, so the client
            renders whatever the profession defines without knowing any of the names.
          items:
            $ref: '#/components/schemas/github_com_bingoring_forin_server_internal_domain_progress.Standing'
          type: array
          uniqueItems: false
        streakCurrent:
          type: integer
        streakLongest:
          type: integer
        xp:
          type: integer
      type: object
    github_com_bingoring_forin_server_internal_domain_progress.Standing:
      properties:
        key:
          type: string
        label:
          type: string
        value:
          type: integer
      type: object
    github_com_bingoring_forin_server_internal_domain_user.Profile:
      properties:
        destination:
          description: e.g. "us"
          type: string
        equippedTitle:
          description: equipped career title id (may be empty)
          type: string
        job:
          description: 'MVP: "nurse"'
          type: string
        nativeLang:
          description: user's language, e.g. "ko"
          type: string
        onboarded:
          description: completed the onboarding wizard
          type: boolean
        targetLang:
          description: language being learned, e.g. "en", "de"
          type: string
        targetLevel:
          description: level in TargetLang, e.g. "B1"
          type: string
        uiLang:
          description: |-
            UILang is the language the SCREENS are drawn in, which is not NativeLang:
            NativeLang tells the AI which language to explain corrections in, so a user
            who wants an English interface must not thereby get English explanations.
            "" means follow NativeLang.
          type: string
        userId:
          type: string
      type: object
    github_com_bingoring_forin_server_internal_domain_user.User:
      properties:
        createdAt:
          type: string
        id:
          description: UUID
          type: string
        status:
          type: string
      type: object
...
    internal_adapters_http.attemptReq:
      properties:
        scenarioId:
          type: string
        score:
          type: integer
      type: object
...
    internal_adapters_http.gradeReq:
      properties:
        grade:
          type: string
      type: object
...
    internal_adapters_http.meResp:
      properties:
        profile:
          $ref: '#/components/schemas/github_com_bingoring_forin_server_internal_domain_user.Profile'
        user:
          $ref: '#/components/schemas/github_com_bingoring_forin_server_internal_domain_user.User'
      type: object
...
    internal_adapters_http.titleReq:
      properties:
        titleId:
          type: string
      type: object
    internal_adapters_http.uiLangReq:
      properties:
        uiLang:
          type: string
      type: object
    internal_adapters_http.updateProfileReq:
      properties:
        destination:
          type: string
        job:
          type: string
        nativeLang:
          type: string
        targetLang:
          type: string
        targetLevel:
          type: string
      type: object
  securitySchemes:
    Bearer:
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: forin server API — Go stdlib, hexagonal. Contract is Go-first (swag
    → openapi → openapi-typescript).
  title: forin API
  version: 0.1.0
openapi: 3.1.0
paths:
  /attempts:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/internal_adapters_http.attemptReq'
              description: scenario clear
              summary: body
        description: scenario clear
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_bingoring_forin_server_internal_domain_progress.Progress'
          description: OK
      security:
      - Bearer: []
      summary: Record a cleared scenario (awards XP, advances streak)
      tags:
      - progress
...
  /me:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_adapters_http.meResp'
          description: OK
      security:
      - Bearer: []
      summary: Current user + profile
      tags:
      - user
...
  /me/curriculum:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  items:
                    $ref: '#/components/schemas/github_com_bingoring_forin_server_internal_curriculum.BuildingGroup'
                  type: array
                type: object
          description: OK
      security:
      - Bearer: []
      summary: Building/floor/curriculum path with per-user progress
      tags:
      - progress
...
  /me/missions:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  items:
                    type: string
                  type: array
                type: object
          description: OK
      security:
      - Bearer: []
      summary: Discovered hidden missions
      tags:
      - progress
  /me/missions/{id}:
    post:
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  items:
                    type: string
                  type: array
                type: object
          description: OK
      security:
      - Bearer: []
      summary: Record a hidden-mission discovery (permanent, idempotent)
      tags:
      - progress
  /me/profile:
    patch:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/internal_adapters_http.updateProfileReq'
              description: onboarding selections
              summary: body
        description: onboarding selections
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_bingoring_forin_server_internal_domain_user.Profile'
          description: OK
      security:
      - Bearer: []
      summary: Save onboarding profile (marks the user onboarded)
      tags:
      - user
  /me/progress:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_bingoring_forin_server_internal_domain_progress.Progress'
          description: OK
      security:
      - Bearer: []
      summary: Current user's growth
      tags:
      - progress
  /me/review:
    get:
      responses: null
      security:
      - Bearer: []
      summary: Review cards due today
      tags:
      - review
  /me/review/{id}/grade:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/internal_adapters_http.gradeReq'
              description: 'grade: again|hard|good|easy'
              summary: body
        description: 'grade: again|hard|good|easy'
        required: true
      responses: null
      security:
      - Bearer: []
      summary: Grade a review card (SM-2 spaced repetition)
      tags:
      - review
...
  /me/stats:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_bingoring_forin_server_internal_domain_progress.GrowthStats'
          description: OK
      security:
      - Bearer: []
      summary: Growth-report aggregates (scenarios, cards, conversation time, attendance)
      tags:
      - progress
  /me/title:
    patch:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/internal_adapters_http.titleReq'
              description: title id (” to un-equip)
              summary: body
        description: title id (” to un-equip)
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_bingoring_forin_server_internal_domain_user.Profile'
          description: OK
      security:
      - Bearer: []
      summary: Equip a career title
      tags:
      - user
  /me/ui-lang:
    patch:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/internal_adapters_http.uiLangReq'
              description: locale code, or \
              summary: body
        description: locale code, or \
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_bingoring_forin_server_internal_domain_user.Profile'
          description: OK
      security:
      - Bearer: []
      summary: Set the app's display language
      tags:
      - user
...